# backend/app/tests/snapshots/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Insight Sphere Backend",
    "description": "API for managing analytical datasets, providing upload/extraction capabilities with strong validation, observability, and documentation.",
    "contact": {
      "name": "Insight Sphere Team",
      "url": "https://github.com/insight-sphere"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/visualization/list": {
      "get": {
        "summary": "List Visualizations Endpoint",
        "operationId": "list_visualizations_endpoint_api_v1_visualization_list_get",
        "parameters": [
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "-created_at",
              "title": "Order By"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы",
              "default": 1,
              "title": "Page",
              "example": 2
            },
            "description": "Номер страницы",
            "examples": {
              "next": {
                "summary": "Следующая страница",
                "value": 2
              },
              "first": {
                "summary": "Первая страница",
                "value": 1
              }
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Количество элементов на странице",
              "default": 20,
              "title": "Page Size",
              "example": 50
            },
            "description": "Количество элементов на странице",
            "examples": {
              "default": {
                "summary": "Стандартное значение",
                "value": 20
              },
              "extended": {
                "summary": "Больше объектов",
                "value": 100
              }
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по названию, описанию и тегам",
              "title": "Search",
              "example": "sales quarterly"
            },
            "description": "Поиск по названию, описанию и тегам"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по тегам",
              "title": "Tags"
            },
            "description": "Фильтр по тегам",
            "examples": {
              "single": {
                "summary": "Один тег",
                "value": [
                  "finance"
                ]
              },
              "multi": {
                "summary": "Несколько тегов",
                "value": [
                  "finance",
                  "forecasting"
                ]
              }
            }
          },
          {
            "name": "types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по типу визуализации",
              "title": "Types"
            },
            "description": "Фильтр по типу визуализации",
            "examples": {
              "chart": {
                "summary": "Тип визуализации",
                "value": [
                  "chart"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visualization/create": {
      "post": {
        "summary": "Create Visualization Endpoint",
        "operationId": "create_visualization_endpoint_api_v1_visualization_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualizationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Visualization Endpoint Api V1 Visualization Create Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visualization/{viz_id}": {
      "get": {
        "summary": "Get Visualization Endpoint",
        "operationId": "get_visualization_endpoint_api_v1_visualization__viz_id__get",
        "parameters": [
          {
            "name": "viz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Viz Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Visualization Endpoint Api V1 Visualization  Viz Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Visualization Endpoint",
        "operationId": "update_visualization_endpoint_api_v1_visualization__viz_id__put",
        "parameters": [
          {
            "name": "viz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Viz Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualizationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Visualization Endpoint Api V1 Visualization  Viz Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Visualization Endpoint",
        "operationId": "delete_visualization_endpoint_api_v1_visualization__viz_id__delete",
        "parameters": [
          {
            "name": "viz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Viz Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Visualization Endpoint Api V1 Visualization  Viz Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visualization/filter": {
      "post": {
        "summary": "Filter Visualizations Endpoint",
        "operationId": "filter_visualizations_endpoint_api_v1_visualization_filter_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualizationFilterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/visualization/create": {
      "post": {
        "summary": "Create Visualization Endpoint",
        "operationId": "create_visualization_endpoint_api_visualization_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualizationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Visualization Endpoint Api Visualization Create Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/visualization/filter": {
      "post": {
        "summary": "Filter Visualizations Endpoint",
        "operationId": "filter_visualizations_endpoint_api_visualization_filter_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualizationFilterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VisualizationCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "chart"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "x_axis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Axis"
          },
          "y_axis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Axis"
          },
          "z_axis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z Axis"
          },
          "insights": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Insights"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "VisualizationCreate"
      },
      "VisualizationFilterRequest": {
        "properties": {
          "filters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters"
          },
          "order_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order By",
            "default": "-created_at"
          }
        },
        "type": "object",
        "title": "VisualizationFilterRequest"
      },
...
    }
  }
}